refactor(conformance): move contract-field inheritance walk to a neutral module
_resolve_contract_fields and _collect_entrypoint_contract_names were
embedded in the B-series deprecation/_contract_compat.py, but this
logic is series-agnostic: the planned K006 (K-series, contract-toolkit)
check needs the same entrypoint-contract and inherited-field resolution
and would otherwise have to import from an unrelated rule series.

Move the entrypoint-contract discovery, field extraction, and
inheritance-walk (plus the canonical-type normalization it depends on)
into a new neutral suite/checks/_contract_fields.py, alongside the
existing convention of top-level shared modules under suite/checks/
(_ast_common, prescriptions helpers, etc.). Move the SDK contract-base
field registry (_sdk_contract_mixins.py) alongside it for the same
reason: it's a dependency of the now-neutral resolver, not something
specific to B005/B006's ledger-tracking concept.

deprecation/_contract_compat.py now only owns the B005/B006
finding-emission logic (comparing resolved fields against the ledger);
the ledger generator and the checker both import the shared resolver
from its new home. Pure refactor: contract_schema.lock.json and the
rendered rule docs are unchanged (--check passes), and the B-series
gate still reports zero B005/B006 findings on this repo.
